+++ src/interfaces.ts
@@ -11,6 +11,8 @@ export interface LoaderConfig {
     compiler?: string;
     configFileName?: string;
     configFileContent?: string;
+    forceIsolatedModules?: boolean;
+    errorsAsWarnings?: boolean;
     transpileOnly?: boolean;
     ignoreDiagnostics?: number[];
     compilerOptions?: ts.CompilerOptions;